Remove unused files including constants.py and hello.py, and restructure the run_evaluation script to import modules from the src directory. Add new modules for data loading, evaluation, and synthesis, including CustomDataset, TrainTestSplit, and PromptEvaluator. Enhance the synthesizer with data generation capabilities and model definitions.
