--- constants.py
@@ -1,1 +1,0 @@
-NUM_REPHRASED_SENTENCES = 6
